+++ modules/School_Setup/MarkingPeriods.php
@@ -67,8 +67,6 @@ if ( ! empty( $_POST['tables'] )
 
 		if ( $id !== 'new' )
 		{
-			$sql = "UPDATE school_marking_periods SET ";
-
 			foreach ( (array) $columns as $column => $value )
 			{
 				if ( $column === 'START_DATE'
@@ -127,13 +125,15 @@ if ( ! empty( $_POST['tables'] )
 						break 2;
 					}
 				}
-
-				$sql .= DBEscapeIdentifier( $column ) . "='" . $value . "',";
 			}
 
-			$sql = mb_substr( $sql, 0, -1 ) . " WHERE MARKING_PERIOD_ID='" . (int) $id . "'";
-
 			$go = true;
+
+			$sql = DBUpdateSQL(
+				'school_marking_periods',
+				$columns,
+				[ 'MARKING_PERIOD_ID' => (int) $id ]
+			);
 		}
 
 		// New: check for Title.