+++ CHANGES.md
@@ -31,6 +31,7 @@ Changes in 11.0
 - Add Insert into or Update DB functions in DBUpsert.php
 - Use DBUpsert() function in Template.fnc.php, Config.fnc.php & StudentAssignments.fnc.php
 - Use DBInsertSQL() function in AddActivity.php, Scheduler.php, Calendar.php, MarkingPeriods.php, MassAssignFees.php & MassAssignPayments.php
+- Use DBUpdateSQL() function in MarkingPeriods.php
 
 Changes in 10.9.6
 -----------------